requirements.txt
flask>=2.3.0
ffmpeg-python>=0.2.0
openai-whisper>=20230314
torch>=2.0.1
werkzeug>=2.3.0
